Add files via upload
BISECTION METHOD
The bisection method is an approximation method to find the roots of the given equation by repeatedly dividing the interval. This method will divide the interval until the resulting interval is found, which is extremely small. Maplesoft can numerically approximate the real roots of an expression using the bisection method.