Add files via upload
NEWTON-RAPHSON METHOD
The Newton-Raphson method (also known as Newton's method) is a way to quickly find a good approximation for the root of a real-valued function f(x) = 0 f(x) = 0 f(x) =0. It uses the idea that a continuous and differentiable function can be approximated by a straight line tangent to it.	Maplesoft uses the Newton-Raphson method to numerically solve some equations. The formulae is given as 
 b = xi – (f(xi)/f'(xi))