Add files via upload
 TAYLOR’S SERIES METHOD 
f(x) = f(a)+(x-a)fʹ(a)+ +…+ +En(x) (𝑥−𝑎) 2 𝑓ʹʹ(𝑎) 2! (𝑥−𝑎) 𝑛 𝑓ʹʹ(𝑎) 2! 
Where 𝐸 𝑛 = (𝑥−𝑎) 𝑛+1 (𝑛+1)! 𝑓 𝑛+1 (ϵ 1 ) 
Taylor’s methods are options within the Maple command Initial Value Problem. To obtain Taylor’s method for the problem, fill in the arguments are the Initial Value Problem command. The options within the Initial Value Problem command are the specification of the first order differential equation to be solved, the initial condition, the final value of the independent variable, the choice of method, the number of steps used to determine the answer(number of steps is determined by subtracting the initial value of x from the final value of x and dividing it by the interval length), the specification of form of the output, and the number of digits of rounding to be used in the computation. Maple responds with an array of data. The output brings a table that gives the values of c, actual solution values y(x), the Taylor approximations and the absolute errors.